backend/requirements.txt
Django==2.2.3
djangorestframework==3.10.1
pymongo==3.8.0
django-cors-headers==3.0.2
